+++ src/redteam/extraction/entities.ts
@@ -1,10 +1,9 @@
 import dedent from 'dedent';
-import { fetchWithCache } from '../../cache';
 import { getEnvBool } from '../../envars';
 import logger from '../../logger';
-import { REQUEST_TIMEOUT_MS } from '../../providers/shared';
 import type { ApiProvider } from '../../types';
-import { RedTeamGenerationResponse, REMOTE_GENERATION_URL } from './common';
+import type { RedTeamTask } from './common';
+import { fetchRemoteGeneration } from './common';
 import { callExtraction, formatPrompts } from './util';
 
 export async function extractEntities(provider: ApiProvider, prompts: string[]): Promise<string[]> {
@@ -12,24 +11,8 @@ export async function extractEntities(provider: ApiProvider, prompts: string[]):
 
   if (useRemoteGeneration) {
     try {
-      const body = {
-        task: 'entities',
-        prompts,
-      };
-
-      const response = await fetchWithCache(
-        REMOTE_GENERATION_URL,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(body),
-        },
-        REQUEST_TIMEOUT_MS,
-        'json',
-      );
-
-      const parsedResponse = RedTeamGenerationResponse.parse(response.data);
-      return parsedResponse.result as string[];
+      const result = await fetchRemoteGeneration('entities' as RedTeamTask, prompts);
+      return result as string[];
     } catch (error) {
       logger.warn(`Error using remote generation, falling back to local extraction: ${error}`);
     }
@@ -44,7 +27,7 @@ export async function extractEntities(provider: ApiProvider, prompts: string[]):
     Each line in your response must begin with the string "Entity:".
   `;
 
-  return await callExtraction(provider, prompt, (output: string) => {
+  return callExtraction(provider, prompt, (output: string) => {
     const entities = output
       .split('\n')
       .filter((line) => line.trim().startsWith('Entity:'))
